+++ src/app/(without-side)/layout.tsx
@@ -1,4 +1,4 @@
-import WhiteBox from '@/components/layouts/box-layout';
+import WhiteBox from '@/components/layouts/without-side/white-box';
 
 export default function WithoutSideLayout({ children }: { children: React.ReactNode }) {
 	return <WhiteBox>{children}</WhiteBox>;
