+++ src/components/layouts/without-side/white-box.tsx
